Feature #28: addressing comments * Corrected path for source code * Improved README.md * pom.xml formatting * Tiny code improvements
